Relocating functions into canopy_functions and renaming imports
