+++ pyrealm/demography/t_model_functions.py
@@ -143,88 +143,3 @@ def calculate_sapwood_masses(
     """
 
     return crown_area * rho_s * height * (1 - crown_fraction / 2) / ca_ratio
-
-
-def calculate_canopy_q_m(m: float, n: float) -> float:
-    """Calculate a q_m value.
-
-    The value of q_m is a constant canopy scaling parameter derived from the ``m`` and
-    ``n`` attributes defined for a plant functional type.
-
-    Args:
-        m: Canopy shape parameter
-        n: Canopy shape parameter
-    """
-    return (
-        m
-        * n
-        * ((n - 1) / (m * n - 1)) ** (1 - 1 / n)
-        * (((m - 1) * n) / (m * n - 1)) ** (m - 1)
-    )
-
-
-def calculate_canopy_z_max_proportion(m: float, n: float) -> float:
-    r"""Calculate the z_m proportion.
-
-    The z_m proportion (:math:`p_{zm}`) is the constant proportion of stem height at
-    which the maximum crown radius is found for a given plant functional type.
-
-    .. math::
-
-        p_{zm} = \left(\dfrac{n-1}{m n -1}\right)^ {\tfrac{1}{n}}
-
-    Args:
-        m: Canopy shape parameter
-        n: Canopy shape parameter
-    """
-
-    return ((n - 1) / (m * n - 1)) ** (1 / n)
-
-
-def calculate_canopy_z_max(z_max_prop: Series, height: Series) -> Series:
-    r"""Calculate height of maximum crown radius.
-
-    The height of the maximum crown radius (:math:`z_m`) is derived from the canopy
-    shape parameters (:math:`m,n`) and the resulting fixed proportion (:math:`p_{zm}`)
-    for plant functional types. These shape parameters are defined as part of the
-    extension of the T Model presented by :cite:t:`joshi:2022a`.
-
-    The value :math:`z_m` is the height above ground where the largest canopy radius is
-    found, given the proportion and the estimated stem height (:math:`H`) of
-    individuals.
-
-    .. math::
-
-        z_m = p_{zm} H
-
-    Args:
-        z_max_prop: Canopy shape parameter of the PFT
-        height: Crown area of individuals
-    """
-    """Calculate z_m, the height of maximum crown radius."""
-
-    return height * z_max_prop
-
-
-def calculate_canopy_r0(q_m: Series, crown_area: Series) -> Series:
-    r"""Calculate scaling factor for height of maximum crown radius.
-
-    This scaling factor (:math:`r_0`) is derived from the canopy shape parameters
-    (:math:`m,n,q_m`) for plant functional types and the estimated crown area
-    (:math:`A_c`) of individuals. The shape parameters are defined as part of the
-    extension of the T Model presented by :cite:t:`joshi:2022a` and :math:`r_0` is used
-    to scale the crown area such that the crown area at the  maximum crown radius fits
-    the expectations of the T Model.
-
-    .. math::
-
-        r_0 = 1/q_m  \sqrt{A_c / \pi}
-
-    Args:
-        q_m: Canopy shape parameter of the PFT
-        crown_area: Crown area of individuals
-    """
-    # Scaling factor to give expected A_c (crown area) at
-    # z_m (height of maximum crown radius)
-
-    return 1 / q_m * np.sqrt(crown_area / np.pi)
